// src/app/enums/raceEnum.ts
export enum RaceEnum {
  ALTMER = "Высокий эльф",
  ARGONIANIN = "Аргонианин",
  BOSMER = "Лесной эльф",
  BRETON = "Бретонец",
  DUNMER = "Данмер",
  IMPEREC = "Имперец",
  KHADJIT = "Каджит",
  NORD = "Норд",
  ORC = "Орк",
  REDGUARD = "Редгард"
}